refactor(utils): move mana_query + search_filters into search_service
mana_query.py and search_filters.py are imported only by search_service
modules (and one test each). Colocate them as internal mixins of the
search_service package; rename search_filters.py to mana_filters.py to
distinguish it from the existing filtering.py mixin (predicates vs.
combined pipeline).

Refs #441
